// src/api/model/apiStatsGet200Response.ts
/**
 * Ulanzi AWTRIX API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiStatsGet200Response { 
    bat: number;
    bat_raw: number;
    type: number;
    lux: number;
    ldr_raw: number;
    ram: number;
    bri: number;
    temp: number;
    hum: number;
    uptime: number;
    wifi_signal: number;
    messages: number;
    version: string;
    indicator1: boolean;
    indicator2: boolean;
    indicator3: boolean;
    app: string;
    uid: string;
}